refactor: 프로젝트 구조 수정 InputForm관련 구조 수정
