+++ components/inputForm/inputFormItems/amount.js
