--- js/listener/paymentMethod.js
@@ -1,3 +1,0 @@
-export function initPaymentMethodListener() {
-    
-}
